requirements.txt
absl-py==1.0.0
accumulation-tree==0.6
alembic==1.0.1
aniso8601==3.0.2
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
bcrypt==3.1.4
beautifulsoup4==4.9.3
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.18.51
botocore==1.21.51
bz2file==0.98
cachetools==2.1.0
certifi==2022.12.7
cffi==1.11.5
chardet==3.0.4
charset-normalizer==2.0.4
Click==7.0
coverage==6.2
cssselect==1.1.0
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.8.2
docutils==0.14
elasticsearch==7.14.0
fastapi==0.68.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.2.1
flask-restx==1.0.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
gast==0.3.3
google-api-core==1.16.0
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
google-cloud-vision==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
h11==0.12.0
h5py==2.10.0
idna==2.7
igraph==0.9.8
ImageHash==4.0
importlib-metadata==4.6.4
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.11.3
jmespath==0.9.3
joblib==1.2.0
json-logging==1.2.0
jsonschema==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
# lxml==4.6.3
Mako==1.2.2
Markdown==3.2.1
MarkupSafe==1.1.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
newspaper3k==0.2.8
nltk==3.6.2
numpy
oauthlib==3.1.0
opt-einsum==3.2.0
packaging==21.0
pact-python==1.4.5
# Pillow==8.1.1
plac==0.9.6
preshed==2.0.1
protobuf==3.18.3
protobuf3-to-dict==0.1.5
psutil==5.8.0
psycopg2==2.8.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pybrake==0.4.5
pycld3==0.22
pycparser==2.19
pydantic==1.8.2
pydub==0.25.1
PyJWT==2.4.0
pyparsing==2.4.7
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
pyudorandom==1.0.0
PyWavelets==1.1.1
PyYAML==5.4.1
redis>=3.5.0
regex==2018.1.10
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rq==1.10.0
rsa==4.7
s3transfer==0.5.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.2.1
sentence-transformers==2.2.0
sentencepiece==0.1.96
sgmllib3k==1.0.0
six==1.12.0
smart-open==1.7.1
soupsieve==2.2.1
SQLAlchemy==1.3.15
SQLAlchemy-Utils==0.36.1
srsly==1.0.5
starlette==0.14.2
tdigest==0.5.2.2
tenacity==6.1.0
tensorboard==2.9.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-hub==0.8.0
tensorflow-text==2.9.0
termcolor==1.1.0
thinc==7.3.1
threadpoolctl==2.2.0
timeout-decorator==0.5.0
tinysegmenter==0.3
tldextract==3.1.0
tmkpy==0.1.1
tokenizers
toolz==0.9.0
torch==1.9.0
tqdm==4.27.0
transformers==4.6.0
twitter-text-parser==1.0.0
typing-extensions==3.10.0.0
ujson==1.35
urllib3==1.26.7
uvicorn==0.15.0
wasabi==0.8.2
Werkzeug==1.0.1
wrapt==1.11.1
zipp==3.5.0
pyacoustid==1.2.2
